front: move ONE_DAY const to consts
